Parameter Set

ParameterSet objects store parameters (prm objects) and add internal validation checks and methods for:

Getting and setting parameter values

Transforming parameter values

Providing dependencies of parameters on each other

Tagging parameters, which may enable further properties

Storing subsets of parameters under prefixes

Active bindings

tags

None -> named_list()
Get tags from the parameter set.

ids

None -> character()
Get ids from the parameter set.

length

None -> integer(1)
Get the length of the parameter set as the number of parameters.

deps

None -> data.table::data.table Get parameter dependencies, NULL if none.

supports

None -> named_list()
Get supports from the parameter set.

tag_properties

list() -> self / None -> list()
If x is missing then returns tag properties if any.
If x is not missing then used to tag properties. Currently properties can either be:
i) 'required' - parameters with this tag must have set (non-NULL) values;
ii) 'linked' - parameters with 'linked' tags are dependent on one another and only one can be set (non-NULL at a time);
iii) 'unique' - parameters with this tag must have no duplicated elements, therefore this tag only makes sense for vector parameters.

values

list() -> self / None -> list()
If x is missing then returns the set (non-NULL) values without transformation or filtering; use $get_values for a more sophisticated getter of values.
If x is not missing then used to set values of parameters, which are first checked internally with the $check method before setting the new values.
See examples at end.

trafo

function() -> self / None -> function()
If x is missing then returns a transformation function if previously set, otherwise NULL.
If x is not missing then it should be a function with arguments x and self, which internally correspond to self being the ParameterSet the transformation is being added to, and x <- self$values. The transformation function is automatically called after a call to self$get_values() and is used to transform set values, it should therefore result in a list. If using self$get_values() within the transformation function, make sure to set transform = FALSE to prevent infinite recursion, see examples at end.

Methods

Public methods

ParameterSet$new()

ParameterSet$print()

ParameterSet$get_values()

ParameterSet$add_dep()

ParameterSet$transform()

ParameterSet$rep()

ParameterSet$extract()

ParameterSet$clone()

Method new()

Constructs a ParameterSet object.

Usage

ParameterSet$new(prms = list(), tag_properties = NULL)

Arguments

prms

(list())
List of prm objects. Ids should be unique.

tag_properties

(list())
List of tag properties. Currently support properties are: i) 'required' - parameters with this tag property must be non-NULL; ii) 'linked' - only one parameter in a linked tag group can be non-NULL and the others should be NULL, this only makes sense with an associated trafo; iii) 'unique' - parameters with this tag must have no duplicated elements, only makes sense for vector parameters.

Examples

prms <- list(
 prm("a", Set$new(1), 1, tags = "t1"),
 prm("b", "reals", 1.5, tags = "t1"),
 prm("d", "reals", 2, tags = "t2")
)
ParameterSet$new(prms)

Method print()

Prints the ParameterSet after coercion with as.data.table.ParameterSet.

Usage

ParameterSet$print(sort = TRUE)

Arguments

sort

(logical(1))
If TRUE (default) sorts the ParameterSet alphabetically by id.

Examples

prms <- list(
 prm("a", Set$new(1), 1, tags = "t1"),
 prm("b", "reals", 1.5, tags = "t1"),
 prm("d", "reals", 2, tags = "t2")
)
p <- ParameterSet$new(prms)
p$print()
print(p)
p

Method get_values()

Usage

ParameterSet$get_values(
  id = NULL,
  tags = NULL,
  transform = TRUE,
  inc_null = TRUE,
  simplify = TRUE
)

Method add_dep()

Usage

ParameterSet$add_dep(id, on, cnd)

Method transform()

Usage

ParameterSet$transform()

Method rep()

Replicate the ParameterSet with identical parameters. In order to avoid duplicated parameter ids, every id in the ParameterSet is given a prefix in the format prefix__id. In addition, linked tags are also given the same prefix to prevent incorrectly linking parameters.

The primary use-case of this method is to treat the ParameterSet as a collection of identical ParameterSet objects.

Note that this mutates the ParameterSet, if you want to instead create a new object then use rep.ParameterSet instead (or copy and deep clone) first.

Usage

ParameterSet$rep(times, prefix)

Arguments

times

(integer(1))
Numer of times to replicate the ParameterSet.

prefix

(character(1)|character(length(times)))
The prefix to add to ids and linked tags. If length 1 then is internally coerced to paste0(prefix, seq(times)), otherwise the length should be equal to times.

Method extract()

Usage

ParameterSet$extract(id = NULL, tags = NULL, prefix = NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

ParameterSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(set6)

## $value examples
p <- ParameterSet$new(list(prm(id = "a", support = Reals$new())))
p$values$a <- 2
p$values

#> $a
#> [1] 2
#> 

## $trafo examples
p <- ParameterSet$new(list(prm(id = "a", 2, support = Reals$new())))
p$trafo

#> NULL

# simple transformation
p$get_values()

#> [1] 2

p$trafo <- function(x, self) {
 x$a <- exp(x$a)
 x
}
p$get_values()

#> [1] 7.389056

# more complex transformation on tags
p <- ParameterSet$new(
  list(prm(id = "a", 2, support = Reals$new(), tags = "t1"),
       prm(id = "b", 3, support = Reals$new(), tags = "t1"),
       prm(id = "d", 4, support = Reals$new()))
)
# make sure `transform = FALSE` to prevent infinite recursion
p$trafo <- function(x, self) {
 out <- lapply(self$get_values(tags = "t1", transform = FALSE),
               function(.x) 2^.x)
 out <- c(out, list(d = x$d))
 out
}
p$get_values()

#> $a
#> [1] 4
#> 
#> $b
#> [1] 8
#> 
#> $d
#> [1] 4
#> 

## ------------------------------------------------
## Method `ParameterSet$new`
## ------------------------------------------------

prms <- list(
 prm("a", Set$new(1), 1, tags = "t1"),
 prm("b", "reals", 1.5, tags = "t1"),
 prm("d", "reals", 2, tags = "t2")
)
ParameterSet$new(prms)

#>    Id Support Value Tags
#> 1:  a     {1}     1   t1
#> 2:  b       ℝ   1.5   t1
#> 3:  d       ℝ     2   t2

## ------------------------------------------------
## Method `ParameterSet$print`
## ------------------------------------------------

prms <- list(
 prm("a", Set$new(1), 1, tags = "t1"),
 prm("b", "reals", 1.5, tags = "t1"),
 prm("d", "reals", 2, tags = "t2")
)
p <- ParameterSet$new(prms)
p$print()

#>    Id Support Value Tags
#> 1:  a     {1}     1   t1
#> 2:  b       ℝ   1.5   t1
#> 3:  d       ℝ     2   t2

print(p)

#>    Id Support Value Tags
#> 1:  a     {1}     1   t1
#> 2:  b       ℝ   1.5   t1
#> 3:  d       ℝ     2   t2

p

#>    Id Support Value Tags
#> 1:  a     {1}     1   t1
#> 2:  b       ℝ   1.5   t1
#> 3:  d       ℝ     2   t2